backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
stripe==14.3.0
python-dotenv==1.2.1
